requirements.txt
# LLM-RAG-WIKI Production Dependencies
PyYAML==6.0.2
chromadb==0.5.23
sentence-transformers==3.3.1
mcp==1.27.0
pytest==8.3.4
